requirements.txt
beautifulsoup4==4.13.4
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-google-genai==3.1.0
langchain-openai==1.1.12
langchainhub==0.1.21
langchain-postgres==0.0.17
langchain-text-splitters==1.1.1
pypdf==6.9.2
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.1
